requirements.txt
# Framework web
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7

# Base de datos
PyMySQL==1.1.0  # Para MySQL (desarrollo local - opcional)
psycopg2-binary>=2.9.9  # Para PostgreSQL (Supabase) - usa >= para mejor compatibilidad con Windows

# Configuración
python-dotenv==1.0.0

# Generación de PDFs
reportlab==4.0.4

# Procesamiento de imágenes
Pillow==11.3.0
pillow-heif>=1.1.1  # Soporte para formato HEIC/HEIF (fotos de iPhone/iPad)

# Seguridad
bcrypt==4.0.1

# Almacenamiento en la nube (Cloudflare R2)
boto3==1.34.0
botocore==1.34.0